# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mister Moderator API
  description: |-
    API мини-приложения **«Мистер Модератор»** для MAX — помощника старосты студенческой группы:
    напоминания, задания, долги, экзамены, материалы, почта и роли внутри группы.

    ### Авторизация
    Мини-приложение передаёт в каждом запросе заголовок `X-Max-Init-Data` — строку
    `WebApp.initData` из MAX Bridge. Сервер проверяет её HMAC-подпись токеном бота.

    * `STRICT_AUTH=true` — заголовок обязателен и подпись проверяется; иначе ответ `401`.
    * `STRICT_AUTH=false` (демо-режим, по умолчанию) — без заголовка запросы выполняются
      от тестового старосты (`max_user_id = test_user_001`): при `USE_TEST_DATA=true` это
      староста группы `iu7-42b` из `backend/seed/test_data.json`, иначе автоматически создаётся
      «Тест Тестов» в группе `TEST-GROUP-01`. Если заголовок передан, пользователь берётся
      из его поля `user` **без проверки подписи** — так проверяются остальные роли.

    ### Роли
    `starosta` (4) → `zam` (3) → `proforg` (2) → `student` (1). Права каждой роли —
    в ответе `GET /api/roles`. При нехватке прав — `403`.

    ### Формат ошибок
    `{"detail": "<сообщение>"}` или `{"detail": {"error": {"code": "...", "message": "..."}}}`.
    Ошибки валидации тела запроса — `422` в стандартном формате FastAPI.
  version: 1.0.0
paths:
...
  /api/materials:
    get:
      tags:
      - Материалы
      summary: Учебные материалы группы
      operationId: list_materials_api_materials_get
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Material'
                title: Response List Materials Api Materials Get
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Список материалов с автором и относительным временем добавления.
      security:
      - MaxInitData: []
      - {}
    post:
      tags:
      - Материалы
      summary: Добавить материал
      description: 'Право `material.upload`. JSON: `title`, `type` (`pdf`, `video`, `other`) и `url` —
        ссылка на файл. Сами файлы хранятся в MAX или облаке, API хранит метаданные и ссылку.'
      operationId: upload_material_api_materials_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadMaterialRequest'
            example:
              title: Конспект лекций по матанализу
              type: pdf
              url: https://example.com/matan.pdf
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - MaxInitData: []
      - {}
  /api/materials/{material_id}/download:
    get:
      tags:
      - Материалы
      summary: Скачать материал
      description: Если у материала есть прямая ссылка — редирект `307` на неё, иначе — метаданные файла
        в MAX для скачивания через MAX Bridge.
      operationId: download_material_api_materials__material_id__download_get
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema: {}
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '307':
          description: Редирект на прямую ссылку на файл
      security:
      - MaxInitData: []
      - {}
  /api/materials/{material_id}:
    delete:
      tags:
      - Материалы
      summary: Удалить материал
      operationId: delete_material_api_materials__material_id__delete
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `material.delete.any` или `material.delete.own` для своих.
      security:
      - MaxInitData: []
      - {}
...
  /api/exams/{exam_id}/materials:
    post:
      tags:
      - Экзамены
      summary: Прикрепить материал к экзамену
      operationId: add_exam_material_api_exams__exam_id__materials_post
      parameters:
      - name: exam_id
        in: path
        required: true
        schema:
          type: integer
          title: Exam Id
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_add_exam_material_api_exams__exam_id__materials_post'
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamMaterial'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: 'Право `exam.addMaterial`. Форма: `title` и необязательная ссылка `url`.'
      security:
      - MaxInitData: []
      - {}
  /api/exams/{exam_id}/materials/{material_id}:
    delete:
      tags:
      - Экзамены
      summary: Открепить материал от экзамена
      operationId: delete_exam_material_api_exams__exam_id__materials__material_id__delete
      parameters:
      - name: exam_id
        in: path
        required: true
        schema:
          type: integer
          title: Exam Id
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `exam.addMaterial`.
      security:
      - MaxInitData: []
      - {}
...
components:
  schemas:
...
    Body_add_exam_material_api_exams__exam_id__materials_post:
      properties:
        title:
          type: string
          title: Title
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      required:
      - title
      title: Body_add_exam_material_api_exams__exam_id__materials_post
...
    ExamMaterial:
      properties:
        id:
          type: string
          title: Id
        examId:
          type: string
          title: Examid
        title:
          type: string
          title: Title
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        addedBy:
          type: string
          title: Addedby
        addedAt:
          type: string
          title: Addedat
      type: object
      required:
      - id
      - examId
      - title
      - addedBy
      - addedAt
      title: ExamMaterial
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Material:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        author:
          type: string
          title: Author
        type:
          type: string
          enum:
          - pdf
          - video
          - other
          title: Type
        createdAt:
          type: string
          title: Createdat
        maxUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Maxurl
      type: object
      required:
      - id
      - title
      - author
      - type
      - createdAt
      title: Material
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    UploadMaterialRequest:
      properties:
        title:
          type: string
          title: Title
        type:
          type: string
          enum:
          - pdf
          - video
          - other
          title: Type
        url:
          type: string
          title: Url
      type: object
      required:
      - title
      - type
      - url
      title: UploadMaterialRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    HTTPError:
      title: HTTPError
      type: object
      required:
      - detail
      properties:
        detail:
          description: 'Текст ошибки или объект {error: {code, message}}'
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: permission_denied
            message:
              type: string
              example: Нет права roles.assign
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: Строка WebApp.initData из MAX Bridge. Обязательна при STRICT_AUTH=true.
servers:
- url: /
  description: 'Текущий хост (через nginx: https://discreetly-enjoyable-koi.cloudpub.ru/)'
- url: http://localhost:8080
...